Extract "ember-cli-dependency-version-checker" file
